pyproject.toml
[project]
name = "google-adk-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.15.0",
    "groq>=0.31.1",
]
